.requirements.txt
pillow~=12.0.0
pyqt6~=6.10.0